vovózinha, proj finalizado

+++ src/Matriz.java
@@ -1,9 +1,9 @@
 public class Matriz {
     public final int rows; // linhas
     public final int cols; // colunas
-    private final int[][] matriz;
+    private final double[][] matriz;
     // !!! constructor
-    public Matriz(int rows, int cols, int[] elements) {
+    public Matriz(int rows, int cols, double[] elements) {
         if (elements.length != rows * cols) {
             throw new IllegalArgumentException("Número de elementos (" + elements.length + ") não corresponde ao tamanho da matriz (" + (rows*cols) + ").");
         }
@@ -13,7 +13,7 @@ public class Matriz {
         this.rows = rows;
         this.cols = cols;
 
-        matriz = new int[rows][cols];
+        matriz = new double[rows][cols];
         int indice = 0;
 
         for (int i = 0; i < rows; i++) {
@@ -24,14 +24,14 @@ public class Matriz {
         }
     }
     // !!! metodo GET
-    public int get(int i, int j) {
+    public double get(int i, int j) {
         if (i < 0 || i >= rows || j < 0 || j >= cols) {
             throw new IndexOutOfBoundsException("Índices fora dos limites da matriz");
         }
         return matriz[i][j];
     }
     // !!! metodo SET
-    public void set(int i, int j, int value) {
+    public void set(int i, int j, double value) {
         if (i < 0 || i >= rows || j < 0 || j >= cols) {
             throw new IndexOutOfBoundsException("Índices fora dos limites da matriz");
         }
