+++ src/Vector.java
@@ -3,22 +3,22 @@ import java.util.Arrays;
 public class Vector extends Matriz {
     public final int dim;
     // !!! constructor
-    public Vector(int dim, int[] elements) {
+    public Vector(int dim, double[] elements) {
         super(dim, 1, elements);
         this.dim = dim;
     }
     // !!! metodo GET
-    public int get(int i) {
+    public double get(int i) {
         return super.get(i, 0);
     }
     // !!! metodo SET
-    public void set(int i, int value) {
+    public void set(int i, double value) {
         super.set(i, 0, value);
     }
     // !!! metodo MOSTRAR
     @Override
     public String toString() {
-        int[] temp = new int[dim];
+        double[] temp = new double[dim];
         for(int i = 0; i < dim; i++) {
             temp[i] = get(i); // usa o get(i) do Vector
         }
